First attempt to migrate to long double
Precision is good. Especially since we're jumping between the very small
value of G, to the very large values for position--I'm sure the added
precision will come in handy over the millions of iterations. And maybe
(hopefully) add up to a 7% error. I really don't want to have to account
for relativity.